requirements.txt
google-cloud-language==2.0.0
psycopg2==2.8.6
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
filelock==3.0.12
huggingface-hub==0.0.16
HyperPyYAML==1.0.0
idna==3.2
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
nltk==3.6.2
numpy==1.21.2
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
speechbrain==0.5.10
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.62.2
transformers==4.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
